Fix float16 overflow in activation pooling (crashed training on Kaggle CUDA)
extract_activation/extract_activations_batch (probe_v3.py) and the
backend's GuardrailEngine.extract_activation all summed hidden states in
float16 before ever casting to float32. Residual-stream activations have
known large-magnitude outlier dimensions; summing ~256 of them in float16
(max ~65504) can overflow to inf. MPS's float16 math apparently tolerated
it (all local runs today were clean), Kaggle's CUDA float16 didn't --
LogisticRegressionCV.fit failed with "Input X contains infinity" on the
first real Kaggle run.

Fix: cast hidden_states to float32 BEFORE the mean/sum reduction, not
after. Verified locally (still MPS, doesn't reproduce the crash, but
confirms no regression): single-text and batched extraction stay
numerically identical (cosine similarity ~1.0) after the change.

Also added a finite-value guard in train_and_eval_probe that raises a
clear, specific error pointing at the extraction code if this ever
recurs, instead of sklearn's generic "contains infinity" traceback with
no indication of where the bad values came from.

+++ backend/app/guardrail_core/inference.py
@@ -77,9 +77,14 @@ class GuardrailEngine:
         base = self.model.model if hasattr(self.model, "model") else self.model
         with torch.no_grad():
             outputs = base(**inputs, output_hidden_states=True)
-        hidden = outputs.hidden_states[self.layer]
+        # float32 before the reduction, not after -- summing ~256 float16
+        # residual-stream values (which have known large-magnitude outlier
+        # dimensions) can overflow to inf. Confirmed on Kaggle's CUDA float16
+        # math during training; fixed at the source here too since this is
+        # the same pooling operation, just at inference time.
+        hidden = outputs.hidden_states[self.layer].float()
         pooled = hidden.mean(dim=1).squeeze(0)
-        return pooled.float().cpu().numpy()
+        return pooled.cpu().numpy()
 
     def check(self, text: str) -> dict:
         start = time.perf_counter()
